requirements.txt
requests==2.32.5
python-dotenv==1.2.2
psycopg2==2.9.11